# spec/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Elementor ULTRA MCP — Companion Plugin REST API
  description: >
    FROZEN contract for the elementor-ultra/v1 REST surface exposed by the
    companion WordPress plugin (plugin/elementor-ultra-mcp). Normative source of
    truth is spec/contracts/10-rest-api.md; this file is the machine-readable
    companion. Where they disagree, the Markdown wins. Auth is WordPress
    Application Passwords (HTTP Basic). Every 2xx response is wrapped in
    {success:true,data:<payload>}; every error is {code,message,data:{status,...}}.
  version: 1.0.0
servers:
  - url: "{wpUrl}/wp-json/elementor-ultra/v1"
    variables:
      wpUrl:
        default: https://example.test
        description: WordPress site base URL
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: base64("<wp_user>:<application_password>")
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: object, additionalProperties: true }
    ErrorEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          description: Stable taxonomy code (WP-F05).
        message: { type: string }
        data:
          type: object
          required: [status]
          properties:
            status: { type: integer }
            op_id: { type: string, nullable: true }
            errors:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
            meta: { type: object, additionalProperties: true }
    FieldError:
      type: object
      required: [path, code, message]
      properties:
        path: { type: string, description: "Dotted JSON-pointer-like path into the request body" }
        code: { type: string }
        message: { type: string }
        meta: { type: object, additionalProperties: true }
...
    OpId:
      type: string
      pattern: '^[A-Za-z0-9_.-]{1,64}$'
...
  responses:
    Success:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    Forbidden:
      description: Lacks capability (E_FORBIDDEN_CAP)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Unprocessable:
      description: Semantic validation failure (E_ATOMIC_VALIDATION, E_ID_COLLISION, E_BUDGET_EXCEEDED, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
paths:
...
  /design/fonts/install:
    post:
      summary: Install a non-catalog font face (contract 18 §7-AI S4)
      operationId: installFont
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, family]
              properties:
                source: { type: string, description: 'http(s) URL, data: URI, or raw base64 (woff2/woff/ttf/otf)' }
                family: { type: string }
                weight: { type: string }
                style: { type: string, enum: [normal, italic, oblique] }
                op_id: { $ref: '#/components/schemas/OpId' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...